refactor: to utils independent files
